Add files via upload
The AMR Abundance Calculator is a Python tool designed to merge data from `.mapstat` and `.res` files, calculate antimicrobial resistance (AMR) abundance, and save the results in a CSV format. It extracts individual and total fragment counts from the `.mapstat` file and merges these with `.res` file data to perform the calculations.